requirements.txt
# Core API dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0
starlette>=0.27.0

# Streamlit app dependencies
streamlit>=1.22.0
requests>=2.28.0
pandas>=1.5.0

# Deep learning and machine learning
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.2.2

# Data processing and visualization
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
opencv-python>=4.7.0
albumentations>=1.3.0
tqdm>=4.65.0

# Image processing utilities
scikit-image>=0.20.0
imageio>=2.25.0

# File operations
shutil  # Part of standard library

# System utilities
io  # Part of standard library
os  # Part of standard library
random  # Part of standard library
time  # Part of standard library

# Data structures
collections  # Part of standard library

# Jupyter notebook environment
jupyter>=1.0.0
ipykernel>=6.22.0
ipywidgets>=8.0.0

# For development and testing
pytest>=7.3.1
black>=23.3.0
